Переносит вспомогательные функции в отдельный файл
