+++ src/const.js
@@ -9,18 +9,6 @@ export const Housing = {
   hotel: `Hotel`,
 };
 
-export const getRating = (num) => (num / MAX_RATING) * 100;
-
-export const getMonth = (date) => {
-  return new Date(date).toLocaleString(`ru`, {
-    month: `long`
-  });
-};
-
-export const getYear = (date) => {
-  return new Date(date).getFullYear();
-};
-
 export const PROPTYPES = {
   OFFER: PropTypes.shape({
     "bedrooms": PropTypes.number,
